requirements.txt
absl-py==1.4.0
aiodns==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
ale-py==0.8.1
alpha-vantage==2.3.1
anyio==3.7.1
appdirs==1.4.4
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1670263926556/work
async-timeout==4.0.3
attrs==23.1.0
AutoROM==0.6.1
AutoROM.accept-rom-license==0.6.1
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1687772187254/work
Backtesting==0.3.3
beautifulsoup4==4.12.2
blinker==1.6.2
bokeh==3.2.1
cachetools==5.3.1
ccxt==3.0.59
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
cloudpickle==2.2.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1691044910542/work
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
Cython==0.29.36
debugpy @ file:///C:/b/abs_c0y1fjipt2/croot/debugpy_1690906864587/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
Farama-Notifications==0.0.4
fastjsonschema==2.18.0
filelock==3.9.0
finta==1.3
Flask==2.3.3
fonttools==4.42.1
frozendict==2.3.8
frozenlist==1.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
grpcio==1.57.0
gym-anytrading @ https://github.com/AminHP/gym-anytrading/archive/master.zip#sha256=e3794e6a53a5b38a7a36a046370570207b23cf3168f1ef01d50958cea468ac5d
gym-notices==0.0.8
gym-trading-env==0.3.3
gymnasium==0.29.1
h11==0.14.0
html5lib==1.1
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1688754491823/work
importlib-resources==6.0.1
ipykernel @ file:///D:/bld/ipykernel_1691424574547/work
ipython @ file:///D:/bld/ipython_1685727936079/work
ipywidgets==8.1.0
itsdangerous==2.1.2
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1690896916983/work
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1687700988094/work
jupyter_core @ file:///D:/bld/jupyter_core_1686775880418/work
jupyterlab-widgets==3.0.8
kiwisolver==1.4.4
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mdurl==0.1.2
mplfinance==0.12.10b0
mpmath==1.2.1
multidict==6.0.4
multitasking==0.0.11
nbformat==5.9.2
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
networkx==3.0
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.8.0.76
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas==2.0.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.3.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1690813113769/work
plotly==5.16.1
polygon-api-client==1.12.4
prettytable==3.8.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1688565951714/work
protobuf==4.24.1
psutil @ file:///C:/ci_311_rebuilds/psutil_1679005906571/work
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycares==4.3.0
pycparser==2.21
pyecharts==2.0.4
pygame==2.5.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1691408637400/work
pyparsing==3.0.9
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2023.3
pywin32==305.1
PyYAML==6.0.1
pyzmq @ file:///C:/b/abs_655zk4a3s8/croot/pyzmq_1686601465034/work
QuantStats==0.0.62
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
Shimmy==1.1.0
simplejson==3.19.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sklearn==0.0.post7
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
stable-baselines3==2.1.0
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tornado @ file:///C:/b/abs_61jhmrrua1/croot/tornado_1690848767317/work
tqdm==4.66.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1688315532570/work
tzdata==2023.3
urllib3==1.26.13
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
webencodings==0.5.1
websocket-client==1.6.2
websockets==11.0.3
Werkzeug==2.3.7
widgetsnbextension==4.0.8
xyzservices==2023.7.0
yarl==1.9.2
yfinance==0.2.28
zigzag==0.3.2
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1689374466814/work
